Let `Moryx.Model` support multiple databases
The `IDatabaseConfigurator` that should be used with a context is not defined
on the context anymore to make it independent of a specific database provider.
`DbContextManager` now sets up contexts based on their configuration while
falling back to `NpgsqlConfigurator` as the default database configurator.

`IDatabaseConfig` has changed by the means of adding `DatabaseConnectionSettings`
and `ConfiguratorTypename` properties. `DatabaseConnectionSettings` combines
the properties, all database providers have in common: A `ConfigurationString`
and `Database` name. If nothing is configured, the `NullModelConfigurator` is
used as default.