Adjust default SQLite connection string to prevent db files being created by default
`Mode=ReadWrite` is added to the default SQLite ConnectionString to prevent the
database being created on a read or write operation, when it doesn't exist.

Databases should be created using the API, to ensure, that they are created
properly. To make sure this happens correctly, the ConnectionString has to be
adjusted for that action.