Use Material UI for database configuration
This also fixes an issue with drop downs being stuck.

For the migrations and setups, the issues with drop downs have been fixed by
removing the drop downs at all and providing a list of available items instead.
Every list item has an 'action button' for the item (migration or setup) to be
executed. So that there is no state anymore to be cached.

Also:

* Correct or 'similar' configurator will be preselected, even on a version
  mismatch
* UI updates properly, when switching between database contexts
* If the database name is `null`, the context name will be used/prefilled as
  the default