Fix test steps in build pipeline
Test projects now reference `coverlet.collector` to create test coverage
reports.

Version 5.8.0 is used since newer versions result in build errors because some
dependency to `Newtonsoft.Json` couldn't be resolved.

Notice, that 5.8 does not exist and gets resolved to 6.0.0 instead. But using
6.0.0 gave the same dependency issues described above.

Providing an output path doesn't seem to work as before, so test files get
copied manually now.

The artifacts plugin was configured to report an error if not test results were
found: No test results found does either mean there are not tests in the
project, so no need to run this step anyway or that for some reason those
results could not be generated which is considered an error