Extend construct method for constructor invocations on resources
Before, we created new resources using reflection and returned them always unsafed. This does not allow resources to act on the resource graph within their constructor methods, however. Hence, we added a conditional invocation to the resource modification interface, which creates a new resource in the resource graph with all required initializations. The behaviour of the endpoint is thereby extended with the option to create such resources with their respective constructors.