Properly respect the queue for action on the job list when adding
While all actions in the JobManager are kept in sync by the _taskQueue, adding a new job through the facade has circumvented the mechanic up to this point. This can lead to undesired behaviour when jobs are added to the job list, while completing jobs are not yet fully removed. We prevent this by enqueueing any additions to the job list, while keeping the external method call synchronized with a task completion source.