`SeamlessScheduler` moves Setups without follow-up to completing
When aborting a clean-up and the production job in turn, the setup is cleared automatically from the jobs in the process engine.
However, the `SeamlessScheduler` did not handle the completing setup in any way.
It is now moved to the completing slots, given that now production job is following and removes it and it's slot on changing to the completed state.