Fix ConnectionString forwarded improperly
Problem:
Using a user other than `postgres` causes the connection test to fail.
This occurs because the ConnectionString lacks a `Database` attribute
during the test.
As a result, the npgsql defaults to using the `User` attribute as the
database name, which leads to a `DatabaseNotFound` error as no matching
database exists.

Solution:
Using a custom config for testing the database server's connection status
with database being "postgres".

INFO:
This migh lead to an edge-case where the user in unable to create/ delete
databases if the "postgres" database has been deleted.

Also added missing `outline` to the save button for the database config.

closes #539