Fix building the node tree
It's common, that OPC UA nodes appear in multiple positions of the tree. But
after they appeared once, they've been skipped for all other occurrences.

This is fixed now: Nodes will be listed, if they occur in different branches of
the tree. They will be ignored though, if they appear more than once on the
same branch. Also, if already found, they won't be updated by following
occurrences, assuming that nodes with the same ID will be the same entity.