Respect simulated activites and movements on stop
When stopping the `ProcessSimulator` we had the issue that movements were not completed and more importantly activites had begun, which did not finish before stopping.
This is especially problematic, when restarting the process engine. The depending simulator module stops first, leaving running activities behind. The process engine then waits in turn for all activities to finish or the timeout. Hence we always ran into the modules timeout.

We solve the issue by canceling all movements on stopping the module and awaiting pending results for simulated activities.
In order to not start new activities while stopping we introduced a `_running` flag, which is evaluated when new activites are ready
(We decided against a complex state machine pattern for this single use case for now).
An additional timeout based on the configured runtimes of the activities makes sure the module stops eventually.