Fix `StepTypeName` migration for products
`ClassName` has to be `Classname`.

Apart from that, the property `string[] Errors` got added to `DatabaseMigrationSummary`
in order to bring errors during migrations to the client.

This also fixes 'deletion' routes in the `DatabaseController`: They didn't
await async calls and thus gave the impression that deletion was successful
even though it threw an exception (after the action returned).